requirements.backend.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
httpx==0.27.0
chromadb==0.5.3
fastembed==0.3.6
pydantic-settings==2.3.4
pydantic[email]==2.8.2
numpy==1.26.4
python-multipart==0.0.9
pdfplumber==0.11.2
PyPDF2==3.0.1
python-docx==1.1.2
reportlab==4.2.2
